move ch stuff to its own file
move ch stuff to its own file

remove extra space

remove last typo
